refactored postUser
