+++ controllers/user/index.js
@@ -1,0 +1,5 @@
+const postUser = require('../../controllers/user/newUser');
+
+module.exports = {
+    postUser
+}
