+++ helpers/index.js
@@ -1,7 +1,11 @@
 const handleError = require('../helpers/handleError')
 const handleReqError = require('../helpers/handleReqError')
+const handleUserRouter = require('../helpers/handleUserRouter')
+const handleConflict = require('../helpers/handleConflict')
 
 module.exports = {
     handleError,
     handleReqError,
+    handleUserRouter,
+    handleConflict
 }
